feat, refactor : 채팅방 리스트 조회의 정렬 기준 변경
기존 채팅방을 나간 시간이 최신인 순 -> 채팅방의 마지막 메시지가 최신인 순
채팅방의 최신 메세지를 반환 100개 -> 1개
