+++ backend/src/main/java/com/example/backend/chat/dto/ChatRoomByMemberDto.java
@@ -13,29 +13,14 @@ public class ChatRoomByMemberDto {
     private String title;
     private List<SimpleMemberProfileDto> joinedMembers;
     private Integer unreadCount;
-    private List<ChatRoom.LastMessage> lastMessages;
+    private ChatRoom.LastMessage lastMessage;
 
 
-    public ChatRoomByMemberDto(ChatRoom chatRoom, Date lastAccessTime, List<SimpleMemberProfileDto> joinedMembers) {
+    public ChatRoomByMemberDto(ChatRoom chatRoom, Integer unreadCount, List<SimpleMemberProfileDto> joinedMembers) {
         this.chatRoomId = chatRoom.getId();
         this.title = chatRoom.getTitle();
         this.joinedMembers = joinedMembers;
-        this.unreadCount = getUnreadCount(chatRoom.getLastMessages(), lastAccessTime);
-        this.lastMessages = chatRoom.getLastMessages();
-    }
-
-    private Integer getUnreadCount(List<ChatRoom.LastMessage> lastMessages, Date lastReadDate) {
-        // binary search - upperbound
-        int lo = 0;
-        int hi = lastMessages.size() - 1;
-        while (lo <= hi) {
-            int mid = (lo + hi) / 2;
-            if (lastMessages.get(mid).getSendTime().after(lastReadDate)) {
-                hi = mid - 1;
-            } else {
-                lo = mid + 1;
-            }
-        }
-        return lastMessages.size() - lo;
+        this.unreadCount = unreadCount;
+        this.lastMessage = chatRoom.getLastMessages().get(chatRoom.getLastMessages().size() - 1);
     }
 }
