+++ backend/src/main/java/com/example/backend/chat/service/ChatRoomListService.java
@@ -12,7 +12,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
-import java.util.Comparator;
 import java.util.Date;
 import java.util.List;
 
@@ -27,16 +26,16 @@ public class ChatRoomListService {
         ChatMember chatMember = chatMemberRepository.findById(memberId);
         List<ChatRoomByMemberDto> chatRoomByMemberDtoList = new ArrayList<>();
 
-        // 마지막 접속 날짜가 가장 최근인 채팅방부터 정렬
-        chatMember.getJoinedChatRooms()
-                .sort(Comparator.comparing(ChatMember.JoinedChatRoom::getLastAccessTime, Comparator.reverseOrder()));
-
         for (ChatMember.JoinedChatRoom joinedChatRoom : chatMember.getJoinedChatRooms()) {
-            ChatRoom chatRoom = chatRoomRepository.findById(joinedChatRoom.getChatRoomId());
+            ChatRoom chatRoom = chatRoomRepository.findByIdWithoutBackupMessages(joinedChatRoom.getChatRoomId());
             Date lastAccessTime = joinedChatRoom.getLastAccessTime();
+            Integer unreadCount = getUnreadCount(chatRoom.getLastMessages(), lastAccessTime);
             chatRoomByMemberDtoList.add(
-                    new ChatRoomByMemberDto(chatRoom, lastAccessTime, getJoinedMembers(chatRoom.getJoinedMemberIds())));
+                    new ChatRoomByMemberDto(chatRoom, unreadCount, getJoinedMembers(chatRoom.getJoinedMemberIds())));
         }
+
+        chatRoomByMemberDtoList.sort((dto1, dto2) ->
+                dto2.getLastMessage().getSendTime().compareTo(dto1.getLastMessage().getSendTime()));
 
         return chatRoomByMemberDtoList;
     }
@@ -48,6 +47,21 @@ public class ChatRoomListService {
             simpleMemberProfileDtoList.add(new SimpleMemberProfileDto(member));
         }
         return simpleMemberProfileDtoList;
+    }
+
+    private Integer getUnreadCount(List<ChatRoom.LastMessage> lastMessages, Date lastReadDate) {
+        // binary search - upperbound
+        int lo = 0;
+        int hi = lastMessages.size() - 1;
+        while (lo <= hi) {
+            int mid = (lo + hi) / 2;
+            if (lastMessages.get(mid).getSendTime().after(lastReadDate)) {
+                hi = mid - 1;
+            } else {
+                lo = mid + 1;
+            }
+        }
+        return lastMessages.size() - lo;
     }
 
 }
